+++ R/throughput_dat.R
@@ -1,5 +1,0 @@
-#' Throughput_dat
-#'
-#' @format A tibble with 25 rows, 4 factors (Day,Operator,Machine,Method), and one response (Throughput)
-#' @source \url{https://www2.isye.gatech.edu/~jeffwu/wuhamadabook/data/throughput.dat}
-"throughput_dat"